+++ src/layout/utils/notifications/index.ts
@@ -1,0 +1,8 @@
+import GlobalNotifications from './GlobalNotifications.svelte';
+import { addNotification, removeNotification } from './api';
+
+export {
+    GlobalNotifications,
+    addNotification,
+    removeNotification
+};