+++ src/layout/utils/notifications/notifications-store.ts
@@ -1,0 +1,5 @@
+import { writable } from 'svelte/store';
+import type { Notification } from './notification';
+
+const notificationsArray: Notification[] = [];
+export let notifications = writable(notificationsArray);