+++ src/shared/components/notification/index.ts
@@ -1,0 +1,7 @@
+import Notification from './Notification.svelte';
+import { NotificationType } from './notification-type';
+
+export {
+    Notification,
+    NotificationType
+};