+++ src/shared/components/notification/notification-type.ts
@@ -1,0 +1,6 @@
+export enum NotificationType {
+    Info,
+    Success,
+    Warning,
+    Error
+}