+++ src/shared/functions/random.ts
@@ -1,0 +1,5 @@
+const generateTimestampHash = (): string => (+new Date).toString(36);
+
+export {
+    generateTimestampHash
+};